+++ frontend/src/components/SelectObj.vue
@@ -1,7 +1,7 @@
 <template>
 <div>
     <div class="meun">
-        <div class="thumbnail">
+        <div class="category thumbnail">
             <div class="key-info clearfix"> <span class="pull-left">
                     <div class="checkbox-wrap"><label><strong>分类</strong></label></div>
                 </span> <span class="pull-right"><i class="fa fa-plus-square" @click="addCategory()" title="点击可以添加更多关键词"></i></span>
@@ -10,6 +10,7 @@
                 <li class="checkbox-wrap" v-for="category in categories" :key="category.id">
                     <label :title="category"><input type="radio" :value="category.id" v-model="checkedCategoryId">{{category.name}}</label>
                     <span class="pull-right">
+                    <i class="fa fa-pencil-square text-gray" @click="editCategory(category)" title="编辑词语"></i>
                         <i class="fa fa-trash text-danger" @click="deleteCategory(category.id)" title="删除词语"></i>
                     </span>
                 </li>
@@ -35,7 +36,7 @@ export default {
       checkedCategoryId: ""
     };
   },
-  props: ["categories"],
+  props: ["categories", "plates"],
   watch: {
     checkedCategoryId(val) {
       this.checkedCategoryId = val;
@@ -49,7 +50,7 @@ export default {
           this.checkedCategoryId = category.id;
         }
       }
-      if (!this.checkedCategoryId) {
+      if (!this.checkedCategoryId || !_.find(val, category => category.id === this.checkedCategoryId)) {
         this.checkedCategoryId = val[0] && val[0].id;
       }
     }
@@ -59,10 +60,8 @@ export default {
       let eventType = JSON.stringify(this.categories) + "addcategory";
       bus.$once(eventType, name => {
         if (this.categories.find(category => category.name === name)) {
-          this.$notify.open({
-            content: "当前名称与已有名称重复！",
-            duration: 1000,
-            type: "danger"
+          this.$modal.alert({
+            content: "当前名称与已有名称重复！"
           });
         } else {
           this.$http
@@ -86,26 +85,51 @@ export default {
     editCategory(category) {
       let eventType = JSON.stringify(this.plate) + "editcategory";
       bus.$once(eventType, newVal => {
-        this.plate.options[newVal] = plate.options[category];
-        delete this.plate.options[category];
-        this.changeData(this.plate);
+        if (this.categories.find(item => item.name === name)) {
+          this.$modal.alert({
+            content: "当前名称与已有名称重复！"
+          });
+        } else {
+          this.$http
+            .post(`/api/category/change`, {
+              id: category.id,
+              name: newVal
+            })
+            .then(() => {
+              bus.$emit("loadPlates");
+            });
+        }
       });
       bus.$emit("open-model", {
         eventType: eventType,
         title: "编辑选项",
         type: "edit",
-        data: category
+        data: category.name
       });
     },
     deleteCategory(id) {
-      this.$modal.confirm({
-        content: "确定删除该分类？",
-        onOk: () =>
-          this.$http.delete(`/api/category/${id}`).then(() => {
-            bus.$emit("loadPlates");
-          })
-      });
+      if (this.plates.filter(item => item.categoryid === id).length) {
+        this.$modal.alert({
+          content: "该分类不为空，禁止删除！"
+        });
+      } else {
+        this.$modal.confirm({
+          content: "确定删除该分类？",
+          onOk: () =>
+            this.$http.delete(`/api/category/${id}`).then(() => {
+              bus.$emit("loadPlates");
+            })
+        });
+      }
     }
   }
 };
 </script>
+<style scoped>
+.category.thumbnail {
+  border: 2px solid #b3b0b0;
+}
+.category .key-info {
+  background-color: #d0ac71;
+}
+</style>
